feat: ⚡ Update IHM with ToF

+++ src/Component/Communication/Decoding/Protocol.cpp
@@ -42,7 +42,7 @@ namespace Component
             return (Core::Status::CORE_ERROR_SIZE);
         }
 
-        uint8_t Protocol::Encode(const Frame &response, char *buffer) {
+        uint16_t Protocol::Encode(const Frame &response, char *buffer) {
             if (buffer == nullptr) {
                 return (0U);
             }
@@ -50,7 +50,7 @@ namespace Component
             const uint8_t cluster = response.GetClusterId();
             const uint8_t command = response.GetCommandId();
 
-            uint8_t       length = snprintf(buffer, 8U, "<%02X%02X%02X", cluster, command, size);
+            uint16_t      length = snprintf(buffer, 8U, "<%02X%02X%02X", cluster, command, size);
 
             for (size_t i = 0U; i < size; i++) {
                 length += snprintf(&buffer[length], 3U, "%02X", response.Get1ByteParam(i));
